fixed the failed test when running on the sqlite db

+++ database/migrations/2020_05_03_154113_rename_column_and_add_columns_in_users_table.php
@@ -13,18 +13,39 @@ class RenameColumnAndAddColumnsInUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('name')->nullable()->change();
-			$table->renameColumn('name', 'first_name');
-			$table->string('last_name')->nullable()->after('name');
-			$table->string('phone')->nullable()->after('email');
-			$table->string('company')->nullable()->after('remember_token');
-			$table->string('address1')->nullable()->after('company');
-			$table->string('address2')->nullable()->after('address1');
-			$table->integer('province_id')->nullable()->after('address2');
-			$table->integer('city_id')->nullable()->after('province_id');
-			$table->integer('postcode')->nullable()->after('city_id');
-        });
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('name')->nullable()->change();
+                $table->renameColumn('name', 'first_name');
+                $table->string('last_name')->nullable()->after('name');
+                $table->string('phone')->nullable()->after('email');
+                $table->string('company')->nullable()->after('remember_token');
+                $table->string('address1')->nullable()->after('company');
+                $table->string('address2')->nullable()->after('address1');
+                $table->integer('province_id')->nullable()->after('address2');
+                $table->integer('city_id')->nullable()->after('province_id');
+                $table->integer('postcode')->nullable()->after('city_id');
+            });
+        } else {
+            Schema::dropIfExists('users');
+            Schema::create('users', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('first_name');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->rememberToken();
+                $table->timestamps();
+                $table->string('last_name')->nullable();
+                $table->string('phone')->nullable();
+                $table->string('company')->nullable();
+                $table->string('address1')->nullable();
+                $table->string('address2')->nullable();
+                $table->integer('province_id')->nullable();
+                $table->integer('city_id')->nullable();
+                $table->integer('postcode')->nullable();
+            });
+        }
     }
 
     /**
@@ -34,16 +55,18 @@ class RenameColumnAndAddColumnsInUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->renameColumn('first_name', 'name');
-			$table->dropColumn('last_name');
-			$table->dropColumn('phone');
-			$table->dropColumn('company');
-			$table->dropColumn('address1');
-			$table->dropColumn('address2');
-			$table->dropColumn('province_id');
-			$table->dropColumn('city_id');
-			$table->dropColumn('postcode');
-        });
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::table('users', function (Blueprint $table) {
+                $table->renameColumn('first_name', 'name');
+                $table->dropColumn('last_name');
+                $table->dropColumn('phone');
+                $table->dropColumn('company');
+                $table->dropColumn('address1');
+                $table->dropColumn('address2');
+                $table->dropColumn('province_id');
+                $table->dropColumn('city_id');
+                $table->dropColumn('postcode');
+            });
+        }
     }
 }
